requirements.txt
pandas
numpy
seaborn
matplotlib
xgboost
scikit-learn
dill
mysql-connector-python
flask
# -e .